Implement asset previewer (#635)

+++ src/building/rotation.c
@@ -64,10 +64,10 @@ static void update_rotation_message(void)
     uint8_t *cursor = data.rotation_text;
     building_type type = building_construction_type();
     cursor += string_from_int(cursor, data.extra_rotation + 1, 0);
-    cursor = string_copy(string_from_ascii("/"), cursor, 100 - (cursor - data.rotation_text));
+    cursor = string_copy(string_from_ascii("/"), cursor, 100 - (int) (cursor - data.rotation_text));
     cursor += string_from_int(cursor, get_num_rotations(type), 0);
-    cursor = string_copy(string_from_ascii(" "), cursor, 100 - (cursor - data.rotation_text));
-    string_copy(lang_get_string(28, type), cursor, 100 - (cursor - data.rotation_text));
+    cursor = string_copy(string_from_ascii(" "), cursor, 100 -  (int) (cursor - data.rotation_text));
+    string_copy(lang_get_string(28, type), cursor, 100 - (int) (cursor - data.rotation_text));
 
     data.warning_id = city_warning_show_custom(data.rotation_text, data.warning_id);
 }
